+++ TelephoneCompany/src/PosPago.java
@@ -4,7 +4,7 @@ import java.util.GregorianCalendar;
 public class PosPago extends Assinante {
 	private float assinatura;
 
-	public PosPago(long cpf, String nome, int numero, float assinatura) {
+	public PosPago(long cpf, String nome, long numero, float assinatura) {
 		super(cpf, nome, numero);
 		this.assinatura = assinatura;
 	}
